+++ cdap-common/src/main/java/co/cask/cdap/common/transaction/MultiThreadTransactionAware.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2016 Cask Data, Inc.
+ * Copyright © 2018 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -14,13 +14,12 @@
  * the License.
  */
 
-package co.cask.cdap.data2.transaction;
+package co.cask.cdap.common.transaction;
 
 import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
 import com.google.common.cache.LoadingCache;
 import com.google.common.cache.RemovalListener;
-import com.google.common.cache.RemovalNotification;
 import org.apache.tephra.Transaction;
 import org.apache.tephra.TransactionAware;
 
@@ -35,13 +34,10 @@ public abstract class MultiThreadTransactionAware<T extends TransactionAware> im
 
   private final LoadingCache<Thread, T> perThreadTransactionAwares =
     CacheBuilder.newBuilder().weakKeys()
-      .removalListener(new RemovalListener<Thread, T>() {
-        @Override
-        public void onRemoval(RemovalNotification<Thread, T> notification) {
-          T value = notification.getValue();
-          if (value != null) {
-            MultiThreadTransactionAware.this.onRemoval(value);
-          }
+      .removalListener((RemovalListener<Thread, T>) notification -> {
+        T value = notification.getValue();
+        if (value != null) {
+          MultiThreadTransactionAware.this.onRemoval(value);
         }
       })
       .build(new CacheLoader<Thread, T>() {
